+++ Sources/RKAPIService/RKAPIService.swift
@@ -1,11 +1,13 @@
 import Foundation
+import Combine
+
 /**
  RKAPIService class. It implements `RKAPIServiceProtocol`. This class serves all the necessary steps to perform a `URLSession` call.
  */
 public class RKAPIService: RKAPIServiceProtocol {
     
-    /// Static instance of `RKAPIService`. It has the default `URLSessionConfiguration`. `URLSessionDelegate` and `OperationQueue` are both nil.
-    public static var shared = RKAPIService(sessionConfiguration: URLSessionConfiguration.default, delegate: nil, queue: nil)
+    /// Static instance of `RKAPIService`. It has ``URLSessionConfiguration.ephemeral`` as configuration. ``URLSessionDelegate`` and ``OperationQueue`` are both nil.
+    public static var shared = RKAPIService(sessionConfiguration: URLSessionConfiguration.ephemeral, delegate: nil, queue: nil)
     
     private let session: URLSession
     
@@ -15,11 +17,15 @@ public class RKAPIService: RKAPIServiceProtocol {
      - Parameters:
         - sessionConfiguration: Receives ``URLSessionConfiguration`` from ``Foundation``
         - delegate: Receives an ``Optional<URLSessionDelegate>`` or ``URLSessionDelegate?`` from ``Foundation``
-        - queue: Receiives an ``Optional<OperationQueue>`` or ``OperationQueue`` from ``Foundation``
+        - queue: Receiives an ``Optional<OperationQueue>`` or ``OperationQueue?`` from ``Foundation``
      */
     public init(sessionConfiguration: URLSessionConfiguration, delegate: URLSessionDelegate?, queue: OperationQueue?) {
         
         session = URLSession(configuration: sessionConfiguration, delegate: delegate, delegateQueue: queue)
+    }
+    
+    public func invalidateAndCancelSession() {
+        session.invalidateAndCancel()
     }
     
     @available(macOS 10.15.0, *)
@@ -40,6 +46,7 @@ public class RKAPIService: RKAPIServiceProtocol {
     }
 }
 
+//MARK: With async/await
 @available(macOS 10.15.0, *)
 @available(iOS 13.0, *)
 extension RKAPIService {
@@ -47,11 +54,11 @@ extension RKAPIService {
      Fetch items with HTTP Get method without any body parameter. Uses async/await concurrency of iOS 13
      
      - Parameters:
-        - urlLink: Receives an `Optional<URL>`
+        - urlLink: Receives an ``Optional<URL>``
      
-     - Throws: An URLError is thrown if urlLink is nil or not a valied URL or server does not provide any response. Also ``HTTPStatusCode`` Error (Custom error) can be thrown if server status code is anything but 200...299
+     - Throws: An ``URLError`` is thrown if urlLink is nil or not a valied URL or server does not provide any response. Also ``HTTPStatusCode`` Error (Custom error) can be thrown if server status code is anything but 200...299
      
-     - Returns: Returns a  ``NetworkResult<T>`` where T is raw Data
+     - Returns: Returns a  ``NetworkResult``
      */
     public func fetchItems(urlLink: URL?) async throws -> NetworkResult<Data> {
         guard let url = urlLink else {
@@ -81,8 +88,9 @@ extension RKAPIService {
     }
     
     /**
-     Fetch items with HTTP method with body parameter. Uses asyn/await method of iOS 13
+     Fetch items with HTTP method.
      
+     Fetch items with HTTP method with body parameter. Uses asyn/await method of iOS 13
      - Parameters:
         - urlLink: Receives an optional URL
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
@@ -91,7 +99,7 @@ extension RKAPIService {
      
      - Throws: An URLError is thrown if urlLink is nil or not a valied URL or server does not provide any response. Also ``HTTPStatusCode`` Error (Custom error) can be thrown if server status code is anything but 200...299
      
-     - Returns: Returns a  ``NetworkResult<T>`` where T is raw Data
+     - Returns: Returns a  ``NetworkResult``
      */
     public func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?) async throws -> NetworkResult<Data> {
         guard let url = urlLink else {
@@ -120,11 +128,90 @@ extension RKAPIService {
             throw URLError(.cannotParseResponse)
         }
         
-        HTTPURLResponse.localizedString(forStatusCode: response.statusCode)
-        
         let status = HTTPStatusCode(rawValue: response.statusCode)
         
         return NetworkResult(data: rawData, response: status)
+    }
+}
+
+//MARK: With Combine Publisher
+@available(macOS 10.15.0, *)
+@available(iOS 13.0, *)
+extension RKAPIService {
+    /**
+     Fetch items with HTTP Get method.
+     
+     Fetch items with HTTP Get method without any body parameter. Uses async/await concurrency of iOS 13.
+     
+     - Parameters:
+        - urlLink: Receives an `Optional<URL>`
+     
+     - Returns: Returns a  ``AnyPublisher<Success, Failure>`` where `Success` is ``NetworkResult`` `Failure` is ``Error``
+     */
+    public func fetchItems(urlLink: URL?) -> AnyPublisher<NetworkResult<Data>, Error> {
+        guard let url = urlLink else {
+            return Fail(error: URLError(.badURL)).eraseToAnyPublisher()
+        }
+        
+        let request = URLRequest(url: url)
+        
+        return session.dataTaskPublisher(for: request)
+            .mapError{ (error) -> URLError in
+                
+                return error
+            }
+            .tryMap{ output in
+                guard let response = output.response as? HTTPURLResponse else {
+                    throw URLError(.cannotParseResponse)
+                }
+                
+                let status = HTTPStatusCode(rawValue: response.statusCode)
+                
+                return NetworkResult(data: output.data, response: status)
+            }
+            .eraseToAnyPublisher()
+    }
+    
+    /**
+     Fetch items with HTTP method.
+     
+     Fetch items with HTTP method with body parameter. Uses asyn/await method of iOS 13.
+     
+     - Parameters:
+        - urlLink: Receives an optional URL
+        - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
+        - body: Optional raw Data for sending to remote server.
+        - jsonData: Accepts a boolean value to determine if HTTP body is in JSON format
+     
+     - Returns: Returns a  ``AnyPublisher<Success, Failure>`` where Success is ``NetworkResult`` Failure is ``Error``
+     */
+    public func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?) -> AnyPublisher<NetworkResult<Data>, Error> {
+        guard let url = urlLink else {
+            return Fail(error: URLError(.badURL)).eraseToAnyPublisher()
+        }
+        var request = URLRequest(url: url)
+        
+        request.httpMethod = httpMethod.rawValue
+        
+        if let valiedBody = body {
+            request.httpBody = valiedBody
+        }
+        
+        return session.dataTaskPublisher(for: request)
+            .mapError{ (error) -> URLError in
+                
+                return error
+            }
+            .tryMap{ output in
+                guard let response = output.response as? HTTPURLResponse else {
+                    throw URLError(.cannotParseResponse)
+                }
+
+                let status = HTTPStatusCode(rawValue: response.statusCode)
+
+                return NetworkResult(data: output.data, response: status)
+            }
+            .eraseToAnyPublisher()
     }
 }
 
@@ -132,11 +219,13 @@ extension RKAPIService {
 @available(macOS 10.10, *)
 extension RKAPIService {
     /**
+     Fetch items with HTTP Get method.
+     
      Fetch items with HTTP Get method without any body parameter. Uses async/await concurrency of iOS 13
      
      - Parameters:
         - urlLink: Receives an `Optional<URL>`
-        - completion: An `@escaping` closure parameter which provides a ``Result<NetworkResult<Data>, Error>`` as return of closure
+        - completion: An `@escaping` closure parameter which provides a ``Result<Success, Failure>`` where `Success` is ``NetworkResult`` and `Failure` is ``Error`` as return of closure
      */
     public func fetchItems(urlLink: URL?, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void ){
         guard let url = urlLink else {
@@ -165,19 +254,19 @@ extension RKAPIService {
                 completion(.success(NetworkResult(data: data, response: status)))
             }
         }
-        
     }
     
     /**
-     Fetch items with HTTP method with body parameter. Uses asyn/await method of iOS 13
+     Fetch items with HTTP method.
+     
+     Fetch items with HTTP method with body parameter. Uses asyn/await method of iOS 13.
      
      - Parameters:
         - urlLink: Receives an optional URL
         - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
         - body: Optional raw Data for sending to remote server.
         - jsonData: Accepts a boolean value to determine if HTTP body is in JSON format
-        - completion: An `@escaping` closure parameter which provides a ``Result<NetworkResult<Data>, Error>`` as return of closure
-     
+        - completion: An `@escaping` closure parameter which provides a ``Result<Success, Failure>`` where `Success` is ``NetworkResult`` and `Failure` is ``Error`` as return of closure
      */
     public func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void ){
         guard let url = urlLink else {