+++ Sources/RKAPIService/RKAPIServiceProtocol.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+import Combine
 
 /**
  Protocol for communication with remote server for data transfer  and receive
@@ -25,7 +26,6 @@ public protocol RKAPIServiceProtocol {
     @available(iOS 13.0, *)
     func fetchItems(urlLink: URL?) async throws -> NetworkResult<Data>
     
-    
     /**
      Fetch items with HTTP Get method without any body parameter. Uses async/await concurrency of iOS 13
      
@@ -36,7 +36,6 @@ public protocol RKAPIServiceProtocol {
     @available(iOS 9.0, *)
     @available(macOS 10.10, *)
     func fetchItems(urlLink: URL?, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void )
-    
     
     /**
      Fetch items with HTTP method with body parameter. Uses asyn/await method of iOS 13
@@ -55,7 +54,6 @@ public protocol RKAPIServiceProtocol {
     @available(iOS 13.0, *)
     func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?) async throws -> NetworkResult<Data>
     
-    
     /**
      Fetch items with HTTP method with body parameter. Uses asyn/await method of iOS 13
      
@@ -69,9 +67,52 @@ public protocol RKAPIServiceProtocol {
     @available(iOS 9.0, *)
     @available(macOS 10.10, *)
     func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void )
+    
+    /**
+     Fetch items with HTTP Get method.
+     
+     Fetch items with HTTP Get method without any body parameter. Uses async/await concurrency of iOS 13.
+     
+     - Parameters:
+        - urlLink: Receives an `Optional<URL>`
+     
+     - Returns: Returns a  ``AnyPublisher<Success, Failure>`` where `Success` is ``NetworkResult`` `Failure` is ``Error``
+     */
+    @available(macOS 10.15.0, *)
+    @available(iOS 13.0, *)
+    func fetchItems(urlLink: URL?) -> AnyPublisher<NetworkResult<Data>, Error>
+    
+    /**
+     Fetch items with HTTP method.
+     
+     Fetch items with HTTP method with body parameter. Uses asyn/await method of iOS 13.
+     
+     - Parameters:
+        - urlLink: Receives an optional URL
+        - httpMethod: ``HTTPMethod`` enum value to send data with that specific method.
+        - body: Optional raw Data for sending to remote server.
+        - jsonData: Accepts a boolean value to determine if HTTP body is in JSON format
+     
+     - Returns: Returns a  ``AnyPublisher<Success, Failure>`` where Success is ``NetworkResult`` Failure is ``Error``
+     */
+    @available(macOS 10.15.0, *)
+    @available(iOS 13.0, *)
+    func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?) -> AnyPublisher<NetworkResult<Data>, Error>
 }
 
 extension RKAPIServiceProtocol {
+    @available(macOS 10.15.0, *)
+    @available(iOS 13.0, *)
+    func fetchItems(urlLink: URL?) -> AnyPublisher<NetworkResult<Data>, Error> {
+        return Fail(error: URLError(.unknown)).eraseToAnyPublisher()
+    }
+    
+    @available(macOS 10.15.0, *)
+    @available(iOS 13.0, *)
+    func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?) -> AnyPublisher<NetworkResult<Data>, Error> {
+        return Fail(error: URLError(.unknown)).eraseToAnyPublisher()
+    }
+    
     func fetchItems(urlLink: URL?, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void ) {}
     
     func fetchItemsByHTTPMethod(urlLink: URL?, httpMethod: HTTPMethod, body: Data?, _ completion: @escaping (Result<NetworkResult<Data>, Error>)-> Void ) {}